fix: fix history bug on non-nerd test
