+++ Main/html/quiz/result.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
     <title>싱숑샹숑</title>
 </head>
-<body>
+<body class="result-page-html-prepared">
     <div id="topFixedBar" class="top-fixed-bar gradient">
         <nav>
             <button class="ranking-icon">
